+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "str-conv" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+- Fix some identifier was misidentified as base64.
+
 ## [1.0.4]
 
 - Fix url match bug.